Refactor & fix sweeper
Fixes #47
