+++ server/src/main/resources/application.properties
@@ -1,3 +1,4 @@
+
 conductor.app.workflow-execution-lock-enabled=true
 conductor.workflow-execution-lock.type=redis
 conductor.redis-lock.serverAddress=redis://localhost:6379
@@ -38,22 +39,25 @@ spring.datasource.hikari.auto-commit=true
 spring.search-datasource.hikari.maximum-pool-size=8
 spring.search-datasource.hikari.auto-commit=true
 
-
-#Background sweeper job
-conductor.workflow-monitor.enabled=true
-#Disable default
-conductor.workflow-reconciler.enabled=false 
-conductor.workflow-repair-service.enabled=false
-
 #System Task Workers
 conductor.app.systemTaskWorkerPollInterval=1
 conductor.app.systemTaskMaxPollCount=10
 conductor.app.systemTaskWorkerThreadCount=10
 
+#Background sweeper job
+#Disable default
+conductor.workflow-reconciler.enabled=false
+conductor.workflow-repair-service.enabled=false
 #Enable the Orkes version
 conductor.orkes.sweeper.enabled=true
-conductor.app.sweeperThreadCount=10
-conductor.sweep-frequency.millis=1
+conductor.orkes.sweeper.frequencyMillis=100
+conductor.orkes.sweeper.sweepBatchSize=10
+conductor.orkes.sweeper.queuePopTimeout=100
+#shares the same sweeper executor
+conductor.app.sweeperThreadCount=16
+
+#Monitor
+conductor.workflow-monitor.enabled=true
 
 #metrics -- only enable what is necessary
 management.endpoints.web.exposure.include=prometheus,health
